backend/leave_management/requirements.txt
aiohttp>=3.8.3,<4.0.0  # Updated version to satisfy langchain
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.8.2
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer>=2.0,<3.0  # Keep the charset-normalizer adjustment
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
distlib==0.3.7
Django==4.2.4
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2024.6.0
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.5
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
npm==0.1.1
numpy==1.26.4
optional-django==0.1.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
Pillow==10.0.0
platformdirs==3.10.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.7
psycopg2-binary==2.9.9
pyarrow==16.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
razorpay==1.4.1
redis==5.0.0
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
sqlparse==0.4.4
streamlit==1.36.0
tenacity==8.4.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
twilio==8.5.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.2
watchdog==4.0.1
yarl==1.9.2
